requirements.txt
playwright==1.40.0
requests==2.31.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-auth==2.23.4
python-dotenv==1.0.0
pandas==2.1.4 